Moved summary chart functions to new visualizer.py file
